+++ deploy_action.rb
@@ -2,7 +2,10 @@
 class DeployAction
   class << self
     def call args, options
-      Remote::Br.update
+      comment = comment(args)
+      task = Context::Br.task_name
+      merge_branches = merge_branches(args)
+      cmds = []
 
       branch = Context::Br.current
       unless branch.include? 'hotfix'
@@ -10,10 +13,7 @@ class DeployAction
         exit
       end
 
-      task = Context::Br.task_name
-      comment = comment(args)
-      merge_branches = merge_branches(args)
-      cmds = []
+      Remote::Br.update
 
       if Context::Code.has_changes?
         cmds << "git add . && git commit -a -m '##{task} #{comment}' #{Remote::Br.exists?(branch) ? " && git pull origin #{branch} " : nil} #{options[:no_push] ? nil : "&& git push origin #{branch}"}"
@@ -45,11 +45,22 @@ class DeployAction
     protected
 
     def merge_branches args
-      args[1..-1].select{|arg| arg != comment(args) }
+      branches = args[1..-1].select{|arg| arg != comment(args) }
+      if branches.size == 0
+        print "Deploy branches not specified." + "\n"
+        print "Try something like $ hf deploy master 'foo' " + "\n"
+        exit
+      end
+      branches
     end
 
     def comment args
-      args.last == 'save' || Context::Br.exists?(args.last) ? nil : args.last
+      comment = args.last == 'deploy' || Context::Br.exists?(args.last) ? nil : args.last
+      if comment.include?('"') || comment.include?("'")
+        print "Invalid comment. Try to avoid ' and \" symbols or fix it in pull request =)" + "\n"
+        exit
+      end
+      comment
     end
 
     def deploy_cmd merge_branch